+++ NEWS.md
@@ -1,5 +1,7 @@
 # tidycovid19 0.0.0.9000
 
+* Included code to scrape information provided by Google COVID-19 Community Mobility Reports
+
 * `*_covid19_spread()`: Added option to exclude unhighlighted countries
 
 * `shiny_covid19_spread()`: Changed API to include all plot options